requirements.txt
httpx>=0.28.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
flask>=2.0.0
